fix(perf): extract homepage into server + client components
- Created HomeContent client component with sticky layout logic
- page.tsx is now a server component that renders HomeContent
- Enables SSR streaming for the homepage shell while keeping
  client-side interactivity (resize observer, sticky columns)
- PRD: docs/prd/prd.codebase-audit.json phase-5
